+++ CHANGELOG.md
@@ -1,7 +1,5 @@
 # TODO
 
-- Add i18n to web editor
-- Metrics for changing PDFs
 - Tooltips
 - Adjust selection algorithm
 - Mobile layout
@@ -12,6 +10,8 @@
 
 ### Added
 - Support for multiple variants in problem texts: `The price of [flour || milk || cheese] has gone up.`
+- Support for stats about what is/isn't changed between reloads of a PDF
+- New leaderboards in stats page showing the above-mentioned stats
 - New `/defaults` endpoint where the frontend queries the backend about the default values for fields, to have a single source of truth
 - Search bar in web editor can now look at more fields